@@ -2,7 +2,7 @@ package com.butul0ve.urbanslang
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.DialogFragment
+import androidx.fragment.app.DialogFragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -12,11 +12,11 @@ import com.butul0ve.urbanslang.utils.SharedPreferencesManager
 const val PRIVACY_POLICY_ACCEPTED = "privacy_policy_accepted_key"
 const val IS_USER_CHOICE = "is_user_choice_key"
 
-class PrivacyPolicyFragmentDialog: DialogFragment() {
+class PrivacyPolicyFragmentDialog: androidx.fragment.app.DialogFragment() {
 
     private lateinit var listener: PrivacyPolicyOnClickListener
 
-    override fun onAttach(context: Context?) {
+    override fun onAttach(context: Context) {
         super.onAttach(context)
 
         try {
@@ -39,13 +39,13 @@ class PrivacyPolicyFragmentDialog: DialogFragment() {
     private fun clickYes() {
         writeToSharedPreferences(true)
         listener.initStatistics()
-        dialog.dismiss()
+        dialog?.dismiss()
     }
 
     private fun clickNo() {
         writeToSharedPreferences(false)
         listener.disableStatistics()
-        dialog.dismiss()
+        dialog?.dismiss()
     }
 
     private fun writeToSharedPreferences(value: Boolean) {
