@@ -1,6 +1,6 @@
 package com.butul0ve.urbanslang.adapter
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -10,7 +10,7 @@ import com.butul0ve.urbanslang.R
 class LetterAdapter(
     private val letters: List<String>,
     private val listener: LetterClickListener
-) : RecyclerView.Adapter<LetterAdapter.LetterViewHolder>() {
+) : androidx.recyclerview.widget.RecyclerView.Adapter<LetterAdapter.LetterViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): LetterViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.letter_item, null)
@@ -26,7 +26,7 @@ class LetterAdapter(
         holder.bind(letter)
     }
 
-    inner class LetterViewHolder(view: View) : RecyclerView.ViewHolder(view), View.OnClickListener {
+    inner class LetterViewHolder(view: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(view), View.OnClickListener {
 
         private val letterButton: Button = view.findViewById(R.id.letter_button)
 
