@@ -3,13 +3,13 @@ package com.butul0ve.urbanslang.mvp.favorites
 import android.app.SearchManager
 import android.content.Context
 import android.os.Bundle
-import android.support.design.widget.FloatingActionButton
-import android.support.design.widget.Snackbar
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.SearchView
-import android.support.v7.widget.Toolbar
+import com.google.android.material.floatingactionbutton.FloatingActionButton
+import com.google.android.material.snackbar.Snackbar
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
+import androidx.appcompat.widget.SearchView
+import androidx.appcompat.widget.Toolbar
 import android.view.*
 import android.widget.ImageView
 import android.widget.TextView
@@ -23,29 +23,23 @@ import javax.inject.Inject
 
 private const val QUERY = "query_extra_key"
 
-class FavoritesFragment : Fragment(), FavoritesMvpView {
+class FavoritesFragment : androidx.fragment.app.Fragment(), FavoritesMvpView {
 
     @Inject
     lateinit var presenter: FavoritesMvpPresenter<FavoritesMvpView>
 
     private lateinit var toolbar: Toolbar
     private lateinit var menuToolbarIcon: ImageView
-    private lateinit var definitionsRV: RecyclerView
+    private lateinit var definitionsRV: androidx.recyclerview.widget.RecyclerView
     private lateinit var noResultTV: TextView
     private lateinit var searchView: SearchView
     private lateinit var deleteFAB: FloatingActionButton
 
-    private lateinit var callback: FragmentCallback
     private lateinit var query: String
 
-    override fun onAttach(context: Context?) {
+    override fun onAttach(context: Context) {
         super.onAttach(context)
         UrbanSlangApp.netComponent.inject(this)
-        try {
-            callback = context as FragmentCallback
-        } catch (ex: ClassCastException) {
-            throw ClassCastException("${activity?.localClassName} must implement FragmentCallback")
-        }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -76,7 +70,7 @@ class FavoritesFragment : Fragment(), FavoritesMvpView {
             query = savedInstanceState.getString(QUERY)!!
         }
 
-        menuToolbarIcon.setOnClickListener { callback.onMenuToolbarClick() }
+//        menuToolbarIcon.setOnClickListener { callback.onMenuToolbarClick() }
         deleteFAB.show()
         deleteFAB.setOnClickListener { showSnackbarClearFavorites() }
         initSearchView()
@@ -118,7 +112,7 @@ class FavoritesFragment : Fragment(), FavoritesMvpView {
     }
 
     override fun onClick(definition: Definition) {
-        callback.onDefinitionClick(definition)
+//        callback.onDefinitionClick(definition)
     }
 
     override fun showSuccessSnackbar() {
