@@ -1,11 +1,10 @@
 package com.butul0ve.urbanslang.mvp.detail
 
-import com.butul0ve.urbanslang.bean.Definition
 import com.butul0ve.urbanslang.mvp.MvpPresenter
 
 interface DetailMvpPresenter<V: DetailMvpView>: MvpPresenter<V> {
 
     fun loadDefinition(id: Long)
 
-    fun handleClick(definition: Definition)
+    fun handleClick(id: Long)
 }
